several changes

+++ run.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env bash
 
-rm resolution_*
-rm exposure_*
-rm tmp/*
+# Check if the folder is not empty; then, remove content
+if [ -n "$(ls -A tmp/)" ]; then
+  rm tmp/*
+fi
 
 gamma_src_code="$(grep -oP '(?<="gamma-src-code": ")[^"]*' $1)"
 output_folder="$(grep -oP '(?<="output": ")[^"]*' $1)"
